Fix: "Cannot log after tests are done" error
This commit fixes the error "Cannot log after tests are done" that was
being thrown whenever the NavBar tests ran. The cause of this error was two things:

 - We never mocked the `useAxios()` call that the BoardCategoryList
   component uses on render, so the test was actually making API calls to the backend which caused a race condition

 - The userEvent calls were not wrapped in `act(() => {})`

This has been fixed by creating a mock for the `useAxios()` hook and by
wrapping the render and userEvent calls in `act(() => {})` functions.